float64

+++ main.py
@@ -205,12 +205,12 @@ def rasterize_geodataframes(
             out_shape=(height, width),
             transform=transform,
             fill=0,
-            dtype='int16',
+            dtype='float64',
             all_touched=True
         )
 
         # Initialize height_plateaus_raster with zeros (float64 for higher precision)
-        height_plateaus_raster = np.zeros((height, width), dtype='float32')
+        height_plateaus_raster = np.zeros((height, width), dtype='float64')
 
 
         # Rasterize each height plateau and use maximum value in case of overlaps
@@ -224,7 +224,7 @@ def rasterize_geodataframes(
                 out_shape=(height, width),
                 transform=transform,
                 fill=0,
-                dtype='int16',
+                dtype='float64',
                 all_touched=True
             )
